Use consistent types in comparisons and printf
- Cast int counts to size_t in loop comparisons with size_t j
- Use %zu format for printing size_t values
- Add (int) casts where size_t j is passed to functions expecting int

+++ virtual.c
@@ -495,11 +495,11 @@ dump_bb(dill_stream c, struct basic_block* bb, int i)
     printf("\n    live_at_end :");
     dump_reg_vec(bb->live_at_end);
     printf("\n    succ :");
-    for (j = 0; j < bb->succ_count; j++) {
+    for (j = 0; j < (size_t)bb->succ_count; j++) {
         printf(" %d", bb->succ_list[j]);
     }
     printf("\n    preds :");
-    for (j = 0; j < bb->pred_count; j++) {
+    for (j = 0; j < (size_t)bb->pred_count; j++) {
         printf(" %d", bb->pred_list[j]);
     }
     if (bb->is_loop_start)
@@ -508,7 +508,7 @@ dump_bb(dill_stream c, struct basic_block* bb, int i)
         printf(" - LOOP_END");
     printf("\n");
     for (j = (size_t)bb->start; j <= (size_t)bb->end; j++) {
-        printf(" %d - ", (int)j);
+        printf(" %zu - ", j);
         virtual_print_insn(
             c, NULL,
             ((char*)c->p->virtual.code_base) + j * sizeof(virtual_insn));
@@ -3081,8 +3081,8 @@ new_emit_insns(dill_stream c,
     }
     init_reg_state(&state, c);
     state.c = c;
-    for (j = 0; j < c->p->vreg_count; j++) {
-        if (dill_type_of(c, 100 + j) == DILL_B) {
+    for (j = 0; j < (size_t)c->p->vreg_count; j++) {
+        if (dill_type_of(c, 100 + (int)j) == DILL_B) {
             /* offset is really size, fix that */
             c->p->vregs[j].offset = dill_localb(c, c->p->vregs[j].offset);
         }
@@ -3091,7 +3091,7 @@ new_emit_insns(dill_stream c,
         basic_block bb = &vmi->bblist[i];
         int insn_count = (int)(bb->end - bb->start);
         int insn_start;
-        for (j = 0; j < c->p->vreg_count; j++) {
+        for (j = 0; j < (size_t)c->p->vreg_count; j++) {
             c->p->vregs[j].assign_loc = -1;
             c->p->vregs[j].in_reg = -1;
             c->p->vregs[j].last_use = -1;
@@ -3099,7 +3099,7 @@ new_emit_insns(dill_stream c,
             c->p->vregs[j].update_in_reg = 0;
             c->p->vregs[j].value_in_mem = -1;
         }
-        for (j = 0; j < c->p->c_param_count; j++) {
+        for (j = 0; j < (size_t)c->p->c_param_count; j++) {
             state.param_info[j].assign_loc = -1;
             state.param_info[j].in_reg = -1;
             state.param_info[j].last_use = -1;
@@ -3111,7 +3111,7 @@ new_emit_insns(dill_stream c,
             virtual_insn* ip = &((virtual_insn*)insns)[j];
             update_vreg_info(&state, bb, ip, (int)j);
         }
-        for (j = 0; j < c->p->vreg_count; j++) {
+        for (j = 0; j < (size_t)c->p->vreg_count; j++) {
             if (get_last_use(&state, (int)j + 100) == -1)
                 continue;
             c->p->vregs[j].use_metric *= insn_count;
@@ -3119,7 +3119,7 @@ new_emit_insns(dill_stream c,
                 (c->p->vregs[j].last_use - c->p->vregs[j].value_in_mem + 1);
             c->p->vregs[j].value_in_mem = bit_vec_is_set(bb->regs_used, (int)j);
         }
-        for (j = 0; j < c->p->c_param_count; j++) {
+        for (j = 0; j < (size_t)c->p->c_param_count; j++) {
             if (c->p->c_param_args[j].is_register) {
                 state.param_info[j].update_in_reg = 0;
                 state.param_info[j].value_in_mem = 0;
